--- src/pages/FighterList.jsx
@@ -1,2 +1,0 @@
-import React { useState } from "react";
-
